# controllers/authCompany_old.controller.js
import jwt from "jsonwebtoken";
import { db } from "../config/connect.js";
import checkEmail from "../middlewares/checkEmail.middleware.js";
import checkPassword from "../middlewares/checkPassword.middleware.js";
import nodemailer from "nodemailer";
import dotenv from "dotenv";
import bcrypt from "bcrypt";
import "express-async-errors";
dotenv.config();

const isProduction = process.env.NODE_ENV === "production";
// Đăng ký
export const register = async (req, res) => {
  try {
    const { nameAdmin, nameCompany, email, password, phone, idProvince, scale, verificationCode } = req.body;

    if (!email) return res.status(409).json("Email không được để rỗng !");
    if (!nameCompany) return res.status(409).json("Tên công ty không được để rỗng !");
    if (!nameAdmin) return res.status(409).json("Tên người đại diện không được để rỗng !");
    if (!checkEmail(email)) return res.status(409).json("Email không hợp lệ.");
    if (!phone || isNaN(phone) || phone?.length > 45) return res.status(409).json("Số điện thoại không hợp lệ !");
    if (!verificationCode) return res.status(409).json("Mã xác thực là bắt buộc !");
    
    if (nameCompany?.length > 255 || nameAdmin?.length > 255 || email?.length > 255 || password?.length > 255)
      return res.status(409).json("Các trường không vượt quá 255 kí tự !");

    if (!checkPassword(password)) return res.status(403).json("Mật khẩu phải bao gồm ít nhất 6 kí tự, trong đó có chữ cái, số, chữ cái viết hoa và kí tự đặt biệt.");

    const promiseDb = db.promise();

    // Kiểm tra email đã tồn tại chưa
    const [existingCompany] = await promiseDb.query(
      'SELECT * FROM companies WHERE email = ?',
      [email]
    );

    if (existingCompany.length > 0) {
      return res.status(409).json("Email đã tồn tại !");
    }

    // Kiểm tra mã xác thực
    const [verification] = await promiseDb.query(
      'SELECT * FROM verification_codes WHERE email = ? AND code = ? AND type = ? AND is_used = FALSE AND expires_at > NOW()',
      [email, verificationCode, 'register']
    );

    if (verification.length === 0) {
      return res.status(400).json({
        success: false,
        message: "Mã xác thực không hợp lệ hoặc đã hết hạn"
      });
    }

    // Hash mật khẩu
    const salt = bcrypt.genSaltSync(10);
    const hashedPassword = bcrypt.hashSync(password, salt);

    // Tạo tài khoản công ty
    await promiseDb.query(
      'INSERT INTO companies (`nameAdmin`, `nameCompany`, `email`, `password`, `phone`, `idProvince`, `scale`) VALUES (?, ?, ?, ?, ?, ?, ?)',
      [nameAdmin, nameCompany, email, hashedPassword, phone, idProvince, scale]
    );

    // Đánh dấu mã đã sử dụng
    await promiseDb.query(
      'UPDATE verification_codes SET is_used = TRUE WHERE id = ?',
      [verification[0].id]
    );

    return res.status(200).json({
      success: true,
      message: "Đăng ký thành công"
    });

  } catch (error) {
    console.error('Error in company register:', error);
    return res.status(500).json({ 
      success: false,
      message: "Lỗi máy chủ", 
      error: error.message 
    });
  }
};

// Đăng nhập
export const login = (req, res) => {
  const { email, password } = req.body;
  const q = "SELECT * FROM companies WHERE email=?";

  if (email && password) {
    if (!checkEmail(email)) return res.status(409).json("Email không hợp lệ.");

    db.query(q, email, (err, data) => {
      if (err) return res.status(500).json(err);
      if (data?.length === 0) return res.status(404).json("Email không tồn tại");
      const checkPassword = bcrypt.compareSync(req.body.password, data[0].password);
      if (!checkPassword) return res.status(401).json("Sai mật khẩu");

      const secret = process.env.MY_SECRET;
      if (!secret) return res.status(500).json("Lỗi cấu hình máy chủ: thiếu MY_SECRET");
      const token = jwt.sign({ id: data[0].id }, secret, { expiresIn: "7d" });
      const { password, ...others } = data[0];

      res
        .cookie("accessToken", token, {
          httpOnly: true,
          sameSite: isProduction ? "none" : "lax",
          secure: isProduction,
          maxAge: 7 * 24 * 60 * 60 * 1000 // 7 ngày
        })
        .status(200)
        .json(others);
    });
  } else {
    return res.status(409).json("Email và mật khẩu không được để rỗng!");
  }
};

//Đăng xuất
export const logout = (req, res) => {
  res
    .clearCookie("accessToken", {
      secure: isProduction,
      sameSite: isProduction ? "none" : "lax",
      path: "/",
    })
    .status(200)
    .json("Đăng xuất thành công");
};

//Quên mật khẩu
export const forgotPassword = (req, res) => {
  const { email } = req.body;

  const q = "SELECT email, nameAdmin, id from companies WHERE email = ?";

  if (!checkEmail(email)) return res.status(403).json("Email không hợp lệ !");

  db.query(q, [email], (err, data) => {
    if (!data?.length) {
      return res.status(403).json("Không tìm thấy email!");
    } else {
      const token = jwt.sign({ id: data[0].id }, process.env.MY_SECRET, { expiresIn: '60s' }); // token tồn tại trong 1 phút

      var transporter = nodemailer.createTransport({
        service: "gmail",
        auth: {
          user: `${process.env.MAIL_NAME}`,
          pass: `${process.env.MAIL_PASSWORD}`,
        },
      });

      const url = `${process.env.URL_REACT}/tao-moi-mat-khau/${data[0].id}/${token}?type=nha-tuyen-dung`;

      const emailHTML = `
      <!DOCTYPE html>
      <html lang="en">
      <head>
        <meta charset="UTF-8">
        <meta name="viewport" content="width=device-width, initial-scale=1.0">
        <title>Xác Nhận Tài Khoản và Tạo Mật Khẩu Mới</title>
      </head>
      <body style="color:#000;font-size:16px">
        <p>Chào nhà tuyển dụng, ${data[0].nameAdmin || "Bạn"},</p>
      
        <p>Chúng tôi nhận được yêu cầu tạo mật khẩu mới cho tài khoản của bạn. Để tiếp tục quá trình này, bạn vui lòng xác nhận yêu cầu bằng cách nhấn vào liên kết dưới đây:</p>
      
        <p><a href="${url}" target="_blank">Liên kết xác nhận tài khoản</a></p>
      
        <p>Lưu ý rằng liên kết này chỉ tồn tại trong vòng 60 giây. Nếu bạn không thực hiện thao tác trong khoảng thời gian này, bạn sẽ cần yêu cầu lại quá trình tạo mật khẩu mới.</p>
      
        <p>Nếu bạn không phải là người yêu cầu thay đổi mật khẩu hoặc có bất kỳ câu hỏi nào khác, vui lòng liên hệ với chúng tôi ngay lập tức qua địa chỉ email <a href="mailto:SDU-JobQuestofficial@gmail.com">SDU-JobQuestofficial@gmail.com</a> của chúng tôi.</p>
      
        <p>Cảm ơn bạn đã sử dụng dịch vụ của chúng tôi.</p>
      
        <p>Trân trọng,<br>
        SDU-JobQuest<br>
        Đội ngũ Hỗ trợ Khách hàng</p>
      </body>
      </html>
      `;

      var mailOptions = {
        from: `${process.env.MAIL_NAME}`,
        to: `${data[0].email}`,
        subject: "SDU-JobQuest || Xác Nhận Tài Khoản và Tạo Mật Khẩu Mới",
        html: emailHTML,
      };

      transporter.sendMail(mailOptions, function (err, info) {
        if (err) {
          console.log(err);
        } else {
          return res.send({ Status: "Success" });
        }
      });
      return res.status(200).json(data[0]);
    }
  });
};

//Tạo mật khẩu mới 
export const resetPassword = (req, res) => {
  const { id, token } = req.params;
  const { password } = req.body;
  
  if (!id || !token || !password) return res.status(403).json("Không tìm thấy!");
  if(!checkPassword(password)) return res.status(403).json("Mật khẩu phải bao gồm ít nhất 6 kí tự, trong đó có chữ cái, số, chữ cái viết hoa và kí tự đặt biệt.");
  if(password?.length > 255) return res.status(409).json("Các trường không vượt quá 255 kí tự!");

  const q = `UPDATE companies SET password = ? WHERE companies.id = ?`;
  const salt = bcrypt.genSaltSync(10);
  const hashedPassword = bcrypt.hashSync(password, salt);

  jwt.verify(token, process.env.MY_SECRET, (err, userInfo) => {
    if (err) {
      return res.status(401).json("Liên kết không đúng hoặc đã quá hạn. Vui lòng xác nhận lại email.");
    } else {
      db.query(q, [hashedPassword, userInfo.id], (err, data) => {
        if (!err) return res.status(200).json("Cập nhật mật khẩu thành công!");
        return res.status(401).json("Lỗi!");
      });
    }
  });
};

//Thay đổi mật khẩu
export const changePassword = (req, res) => {
  const { passwordOld, password } = req.body;
  const id = req.params.id;
  const token = req.cookies?.accessToken;

  if (!id) return res.status(403).json("Không tìm thấy người dùng!");
  if (!token) return res.status(401).json("Chưa đăng nhập !");
  if(!checkPassword(password)) return res.status(403).json("Mật khẩu phải bao gồm ít nhất 6 kí tự, trong đó có chữ cái, số, chữ cái viết hoa và kí tự đặt biệt.");
  if(password?.length > 255) return res.status(409).json("Các trường không vượt quá 255 kí tự !");
  
  const q = "SELECT * FROM companies WHERE id=?";

  db.query(q, id, (err, data) => {
    if (err) return res.status(500).json(err);
    if (data?.length === 0) return res.status(404).json("Nguời dùng không tồn tại");

    const checkPassword = bcrypt.compareSync(passwordOld, data[0].password);

    if (!checkPassword) return res.status(401).json("Mật khẩu cũ không đúng !");

    const q2 = `UPDATE companies SET password = ? WHERE companies.id = ?`;
    const salt = bcrypt.genSaltSync(10);
    const hashedPassword = bcrypt.hashSync(password, salt);

    jwt.verify(token, process.env.MY_SECRET, (err, userInfo) => {
      if (err) {
        return res.status(401).json("Lỗi!");
      } else {
        db.query(q2, [hashedPassword, userInfo.id], (err, data) => {
          if (!err) return res.status(200).json("Cập nhật mật khẩu thành công!");
          return res.status(401).json("Lỗi!");
        });
      }
    });
  });
};